Fix formatting

+++ ESHTTPOperation.m
@@ -523,7 +523,7 @@ static int32_t GetOperationID(void)
 
 - (NSString *)description
 {
-	return [NSString stringWithFormat:@"<%@ : %p>\n{\n\tRequest: %@\n\tResponse: %@\n\tID: %d\n\tError: %@\n}", NSStringFromClass([self class]), self, self.request, self.lastResponse, self.operationID, self.error];
+	return [NSString stringWithFormat:@"<%@ : %p>\n{\n\tRequest: %@\n\tResponse: %@\n\tID: %ld\n\tError: %@\n}", NSStringFromClass([self class]), self, self.request, self.lastResponse, (long)self.operationID, self.error];
 }
 
 - (void)setUploadProgressBlock:(ESHTTPOperationUploadBlock)uploadProgress
